refactor(web/detail): split into components/ and lib/ subfolders
