+++ apps/web/src/components/detail/components/ManualActions.tsx
