+++ apps/web/src/components/detail/components/RightRail.tsx
