+++ apps/web/src/components/detail/components/TransitionButton.tsx
