+++ apps/web/src/components/detail/lib/sections.ts
